[Prover Service] Rename config dir to external_resources and merge JWK fetching.
