+++ prover-service/src/external_resources/prover_config.rs
