+++ prover-service/src/lib.rs
@@ -2,10 +2,9 @@
 
 extern crate core;
 
-pub mod config;
 pub mod error;
+pub mod external_resources;
 pub mod input_processing;
-pub mod jwk_fetching;
 pub mod metrics;
 pub mod request_handler;
 pub mod training_wheels;
