+++ prover-service/src/main.rs
@@ -6,8 +6,8 @@ use aptos_logger::{error, info, warn};
 use clap::Parser;
 use hyper::service::{make_service_fn, service_fn};
 use hyper::Server;
-use prover_service::config::prover_config;
-use prover_service::config::prover_config::ProverServiceConfig;
+use prover_service::external_resources::prover_config::ProverServiceConfig;
+use prover_service::external_resources::{jwk_fetching, prover_config};
 use prover_service::request_handler::prover_state::{ProverServiceState, TrainingWheelsKeyPair};
 use prover_service::request_handler::{deployment_information, handler};
 use prover_service::*;
