+++ prover-service/src/request_handler/handler.rs
@@ -1,6 +1,6 @@
 // Copyright (c) Aptos Foundation
 
-use crate::config::prover_config::ProverServiceConfig;
+use crate::external_resources::prover_config::ProverServiceConfig;
 use crate::request_handler::deployment_information::DeploymentInformation;
 use crate::request_handler::prover_handler;
 use crate::request_handler::prover_state::ProverServiceState;
